requirements/base.txt
pytz==2022.7.1  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.5.1  # https://github.com/redis/redis-py
hiredis==2.2.2  # https://github.com/redis/hiredis-py
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.4.0  # https://github.com/celery/django-celery-beat
flower==1.2.0  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==4.1.7  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.52.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.14.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.26.0  # https://github.com/tfranzel/drf-spectacular


dj-rest-auth @ git+https://github.com/iMerica/dj-rest-auth.git@master
djangorestframework-simplejwt>=5.2.0
coreapi>=2.3.3
django-extensions>=3.2.1
django-braces>=1.15.0
django-extra-views>=0.14.0
django-model-utils>=0.4.3
django-oauth-toolkit>=2.2.0


numpy>=1.24.2
